Improve drawing routine
Due to some tests not included in this commit that involved drawing large flocks, the old routine would create a polygon between each of the boids. I both eliminated that bug and sped up the drawing routine by instead pre-drawing the boids on an offscreen canvas, and then copying them over to the main canvas every frame.